+++ src/DataTable.js
@@ -1,31 +1,34 @@
-import React, { useCallback, useRef } from 'react';
+import React, { useCallback, useRef, useMemo } from 'react';
 
 import "./TrackRowInfoControl.scss";
 import "./DataTable.scss";
-import { CLOSE } from './utils/icons';
 
 /**
  * Component for data table.
  * @prop {number} margin The gap between the inner table and the viewport.
  * @prop {array} rows Array of rows to render data table.
  * @prop {array} columns Array of column names in data table.
+ * @prop {array} selectedRows Array of rows selected by a user.
  * @prop {string} title A title for the table. Optional.
  * @prop {string} subtitle A subtitle for the table. Optional.
  * @prop {boolean} isLoading Whether the data is still loading, in which case show a spinner.
  * @prop {array} expoNotations A list of columns names that need to use exponential notations.
  * @prop {(function|null)} onCheckRows If a function is provided, a checkbox will be shown for each row.
+ * @prop {function} onSelect A function to call upon the selection on rows in a data table.
  * On change of any checkbox elements, an array of all checked row objects will be passed to the function. By default, null.
  */
 export default function DataTable(props) {
     const {
         margin,
         rows = [], columns = [],
+        selectedRows,
         title = "Data Preview",
         subtitle,
         isLoading = false,
         onCheckRows = null,
         expoNotations = [],
-        onClose
+        onClose,
+        onSelect
     } = props;
 
     // Store the currently-checked row indices in a mutable set object.
@@ -52,73 +55,58 @@ export default function DataTable(props) {
         </tr>
     );
 
-    const bodyRows = rows.map((d, i) => {
-        const checkboxCell = (onCheckRows ? (
-            <td>
-                <input
-                    type="checkbox"
-                    name="data-table-checkbox"
-                    value={i}
-                    onChange={handleInputChange}
-                />
-            </td>
-        ) : null);
-        const dataCells = columns.map((c, j) => {
-            return (
-                <td key={j}>
-                    {expoNotations.includes(c) && +d[c] ? Number.parseFloat(d[c]).toExponential(2) : d[c]}
+    const bodyRows = useMemo(() => { 
+        return rows.map((d, i) => {
+            const checkboxCell = (onCheckRows ? (
+                <td>
+                    <input
+                        type="checkbox"
+                        name="data-table-checkbox"
+                        value={i}
+                        onChange={handleInputChange}
+                    />
                 </td>
+            ) : null);
+            const dataCells = columns.map((c, j) => {
+                return (
+                    <td key={j}>
+                        {expoNotations.includes(c) && +d[c] ? Number.parseFloat(d[c]).toExponential(2) : d[c]}
+                    </td>
+                );
+            });
+            return (
+                <tr 
+                    key={i}
+                    className={selectedRows.includes(i) ? 'data-table-row-selected' : 'data-table-row'}
+                    onClick={() => { onSelect([i]) }}
+                >
+                    {checkboxCell}{dataCells}
+                </tr>
             );
-        });
-        return <tr key={i}>{checkboxCell}{dataCells}</tr>;
-    });
+        })
+    }, [selectedRows, expoNotations, onCheckRows]);
 
     return (
-        <div 
-            className="cisvis-data-table-container"
-            style={{ 
-                margin,
-                width: `calc(100% - ${margin * 2}px)`,
-                height: `calc(100% - ${margin * 2}px)`
-            }}>
-            <h4 className="cisvis-table-title">{title}</h4>
-            <span className="cisvis-table-subtitle">
-                {isLoading ? (
-                    <span className="cisvis-progress-ring" />
-                ) : (subtitle ? (
-                    <b>{subtitle}</b>
-                ) : null)}
-            </span>
-            <span style={{ verticalAlign: "middle", display: "inline-block" }}>
-                <svg
-                    className={`chw-button`}
-                    style={{ color: "gray", background: "none" }}
-                    onClick={() => onClose()} 
-                    viewBox={CLOSE.viewBox}
-                >
-                    <title>Close data table</title>
-                    <path d={CLOSE.path} fill="currentColor"/>
-                </svg>
-            </span>
-            <div
-                style={{
-                    height: "calc(100% - 40px)",
-                    overflowY: "auto"
-                }}
-            >
-                {bodyRows ? (
-                    <form>
-                        <table className="cisvis-table">
-                            <thead>
-                                {headRow}
-                            </thead>
-                            <tbody>
-                                {bodyRows}
-                            </tbody>
-                        </table>
-                    </form>
-                ) : null}
-            </div>
+        <div
+            style={{
+                overflowY: "auto",
+                border: "1px solid lightgray",
+                background: "#f9f9f9",
+                padding: '1px'
+            }}
+        >
+            {bodyRows ? (
+                <form>
+                    <table className="cisvis-table">
+                        <thead>
+                            {headRow}
+                        </thead>
+                        <tbody>
+                            {bodyRows}
+                        </tbody>
+                    </table>
+                </form>
+            ) : null}
         </div>
     );
 }