+++ src/TrackRowInfo.js
@@ -9,8 +9,8 @@ import TrackRowInfoVis from "./TrackRowInfoVis.js";
  * @prop {number} trackY The track vertical offset.
  * @prop {number} trackWidth The track width.
  * @prop {number} trackHeight The track height.
- * @prop {object[]} rowInfo Array of JSON objects, one object for each sample, without filtering/sorting based on selected rows.
- * @prop {object[]} transformedRowInfo The `rowInfo` array after transforming by filtering and sorting according to the selected rows.
+ * @prop {object[]} rowInfo The array of JSON Object containing row information.
+ * @prop {object[]} transformedRowInfo The `rowInfo` array after aggregating, filtering, and sorting rows.
  * @prop {array} rowInfoAttributes Array of JSON object, one object for the names and types of each attribute.
  * @prop {string} rowInfoPosition The value of the `rowInfoPosition` option.
  * @prop {object} rowSort The options for sorting rows.
@@ -26,7 +26,7 @@ export default function TrackRowInfo(props) {
 
     const {
         trackX, trackY,
-        trackWidth, trackHeight, 
+        trackWidth, trackHeight,
         rowInfo,
         transformedRowInfo, 
         rowInfoAttributes,