+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 ## 0.4.0 - in progress
 
 ### Added
+- Added row aggregation options (e.g., `rowAggregate`) to aggregate rows based on nominal fields metadata.
+- Added visual indicators (vertical black bars) to show which rows are aggregated.
 
 ### Changed
 - Changed the table view to be shown as a modal view.
 - Updated format of API results (e.g., use consistent letter cases and use scientific notations for long numbers).
+- Filtering information in range sliders and keyword search boxes is based on the aggregated values (`aggregatedRowInfo`), instead of original `rowInfo`.
+- Fixed a SVG export bug which caused visualizations to be exported from previous datasets (corresponding to previous `options` prop values in the demo).
 
 ## 0.3.0 - 05/27/20